Use async script to load Google Fonts (for now)
The concern is because we don't know if the Jekyll blog is going to be
deployed in https (GitHub Pages deployment set default to https) or to
custom domain in which it doesn't know if it has https enabled.

Plus the fact that agnostic-protocol links are bad, so we're using JS to
load it for now, and marking it as async.